app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
attrs==25.4.0
azure-core==1.36.0
azure-eventhub==5.15.0
azure-eventhub-checkpointstoreblob-aio==1.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
frozenlist==1.8.0
idna==3.11
isodate==0.7.2
multidict==6.7.0
paho-mqtt==2.1.0
propcache==0.4.1
pycparser==2.23
python-dotenv==1.2.1
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.9.0
yarl==1.22.0
